The menu class has suffered vast renovation and improvement. It's not done yet, and I can't test it because my computer is showing most of the Gameplay class as containing errors for some reason.  The menus now require that you pass in the Player object, and they proceed to modify the board.  Since all java parameters are passed by reference, i don't need to return the board, per se, even though on success I likely still need to learn how to utilize the network class to send the board to the opponent.
The menu classes will return a 1 on success, a 2 on a cancel, and a 0 for the quit, per the integer return type suggestion i received.  All that is need done is for the StartMenu to be called from somewhere... anywhere really, and have the player object passed into it. (high coupling, dang it)