Added some sound capabilites which don't seem to work on my computer so I'm not sure if it works or not.
Re-updated the readme from the repository reverting that happened before.  I forgot to recommit the changes I made afterwards

Fixed some issues with the boards not being displayed at the right times.

Started to go crazy because it's 12:41 and I still have to finish my 3640 filter assignment.


I think I'm done for the night.  It's been a good day.